Node: Rename wormchain to wormconn
