+++ node/pkg/wormconn/send_tx.go
@@ -1,4 +1,4 @@
-package wormchain
+package wormconn
 
 import (
 	"context"
